+++ R/netmeta-package.R
@@ -45,12 +45,16 @@
 #' \itemize{
 #' \item rankograms (\code{\link{rankogram}}) (Salanti et al., 2011);
 #' \item ranking of treatments (\code{\link{netrank}}) based on
-#'   P-scores (Rücker & Schwarzer, 2015) or the Surface Under the
-#'   Cumulative RAnking curve (SUCRA) (Salanti et al., 2011);
+#'   P-scores (Rücker & Schwarzer, 2015), the Surface Under the
+#'   Cumulative RAnking curve (SUCRA) (Salanti et al., 2011), and
+#'   other ranking metrics;
 #' \item partial order of treatment rankings (\code{\link{netposet}},
 #'   \code{\link{plot.netposet}}) and Hasse diagram
 #'   (\code{\link{hasse}}) according to Carlsen & Bruggemann (2014)
-#'   and Rücker & Schwarzer (2017).
+#'   and Rücker & Schwarzer (2017);
+#' \item VišeKriterijumska Optimizacija I Kompromisno Rešenje (VIKOR)
+#'   multi-criteria decision analysis method (\code{\link{vikor}})
+#'   (Opricovic & Tzeng, 2004).
 #' }
 #' 
 #' Available functions to evaluate network inconsistency:
@@ -189,6 +193,12 @@
 #' Illustrating the assumptions of meta-regression in treatment networks.
 #' Preprint available at \emph{Research Square},
 #' \doi{10.21203/rs.3.rs-8235913/v1}
+#'  
+#' Opricovic S, Tzeng GH (2004):
+#' Compromise solution by MCDM methods: A comparative analysis of VIKOR and
+#' TOPSIS.
+#' \emph{European Journal of Operational Research},
+#' \bold{156}, 445--55
 #' 
 #' Papakonstantinou, T., Nikolakopoulou, A., Rücker, G., Chaimani, A.,
 #' Schwarzer, G., Egger, M., Salanti, G. (2018):
@@ -276,15 +286,17 @@
 #'   
 #' @importFrom methods as
 #'
-#' @importFrom utils installed.packages packageDescription capture.output packageVersion
+#' @importFrom utils installed.packages packageDescription capture.output packageVersion combn
+#' 
+#' @importFrom reshape2 melt
 #'
-#' @importFrom igraph E<- V<- E V all_simple_paths delete_edges get.shortest.paths graph_from_adjacency_matrix graph_from_edgelist gsize set_edge_attr set_vertex_attr head_of tail_of
+#' @importFrom igraph E<- V<- E V all_simple_paths delete_edges get.shortest.paths graph graph_from_adjacency_matrix graph_from_edgelist gsize set_edge_attr set_vertex_attr head_of tail_of
 #'
 #' @importFrom grDevices colours col2rgb heat.colors rainbow rgb xy.coords
 #'
 #' @importFrom graphics axis box lines par points plot polygon rect text strheight strwidth title
 #'
-#' @importFrom ggplot2 ggplot aes xlab ylab labs element_blank element_line element_rect element_text expand_limits geom_col geom_line geom_step geom_tile geom_text ggtitle geom_abline geom_hline geom_point geom_vline scale_fill_gradient2 scale_x_discrete scale_y_discrete scale_x_continuous scale_x_discrete scale_y_discrete theme theme_classic theme_dark scale_color_manual coord_cartesian guide_legend guides margin
+#' @importFrom ggplot2 annotate coord_fixed ggplot aes xlab ylab labs element_blank element_line element_rect element_text expand_limits geom_col geom_line geom_step geom_tile geom_text ggtitle geom_abline geom_hline geom_point geom_vline scale_fill_gradient scale_fill_gradient2 scale_x_discrete scale_y_discrete scale_x_continuous scale_x_discrete scale_y_discrete theme theme_classic theme_dark theme_minimal scale_color_manual scale_y_continuous coord_cartesian guide_legend guides margin
 #' 
 #' @importFrom ggrepel geom_text_repel
 #'
@@ -292,11 +304,11 @@
 #' 
 #' @importFrom grid arrow convertHeight convertWidth convertX convertY drawDetails gpar grid.clip grid.draw grid.lines grid.newpage grid.rect grid.roundrect grid.text grob popViewport pushViewport stringWidth unit viewport
 #' 
-#' @importFrom dplyr %>% filter select rename starts_with relocate last_col mutate if_else bind_rows pull case_when distinct count summarise inner_join
+#' @importFrom dplyr %>% filter select rename starts_with relocate last_col mutate if_else bind_rows pull case_when distinct count summarise inner_join arrange if_any
 #' 
-#' @importFrom stringr str_length
+#' @importFrom stringr str_length str_to_sentence
 #' 
-#' @importFrom tidyr pivot_longer
+#' @importFrom tidyr pivot_longer drop_na
 #' 
 #' @importFrom magrittr %<>%